# lib/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a REST server for ioBroker."
  version: "2.0.3"
  title: "ioBroker Swagger UI"
  contact:
    email: "admin@iobroker.net"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
...
securityDefinitions:
  basicAuth:
    type: basic
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your bearer token in the format **Bearer &lt;token&gt;**"
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /getHistory:
    x-swagger-router-controller: history
    post:
      tags:
        - "history"
      summary: "Reads history"
      operationId: "postHistory"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "Options"
          in: "body"
          description: "Options for history"
          required: true
          schema:
            $ref: "#/definitions/HistoryRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HistoryResponse"
        422:
          description: "Invalid options supplied"
  /getHistory/{stateId}:
    x-swagger-router-controller: history
    get:
      tags:
        - "history"
      summary: "Reads history"
      operationId: "getHistory"
      produces:
        - "application/json"
      parameters:
        - name: "stateId"
          in: "path"
          description: "State ID"
          type: "string"
          required: true
        - name: start
          in: "query"
          type: "number"
          description: "Start time in ms"
        - name: end
          in: "query"
          type: "number"
          description: "End time in ms"
        - name: count
          in: "query"
          type: "number"
          description: "number of values if aggregate is 'onchange' or number of intervals if other aggregate method. Count will be ignored if step is set, else default is 500 if not set"
        - name: from
          in: "query"
          type: "boolean"
          description: "if FROM field should be included in answer"
        - name: ack
          in: "query"
          type: "boolean"
          description: "if ACK field should be included in answer"
        - name: q
          in: "query"
          type: "boolean"
          description: "if Q field should be included in answer"
        - name: addId
          in: "query"
          type: "boolean"
          description: "if ID field should be included in answer"
        - name: limit
          in: "query"
          type: "number"
          description: "do not return more entries than limit"
        - name: ignoreNull
          in: "query"
          type: "boolean"
          description: "if null values should be included (false), replaced by last not null value (true) or replaced with 0 (0)"
        - name: removeBorderValues
          in: "query"
          type: "boolean"
          description: "By default additional border values are returned to optimize charting. Set this option to true if this is not wanted (e.g. for script data processing)"
        - name: returnNewestEntries
          in: "query"
          type: "boolean"
          description: "The returned data are always sorted by timestamp ascending. When using aggregate none and also providing count or limit this means that normally the oldest entries are returned (unless no start data is provided). Set this option to true to get the newest entries instead."
        - name: aggregate
          in: "query"
          type: "string"
          description: "Aggregation method"
          enum: [minmax, max, min, average, total, count, percentile, quantile, integral, none]
        - name: percentile
          in: "query"
          type: "number"
          description: "(optional) when using aggregate method percentile defines the percentile level (0..100)(defaults to 50)"
        - name: quantile
          in: "query"
          type: "number"
          description: "(optional) when using aggregate method quantile defines the quantile level (0..1)(defaults to 0.5)"
        - name: integralUnit
          in: "query"
          type: "string"
          description: "(optional) when using aggregate method integral defines the unit in seconds (defaults to 60s). e.g. to get integral in hours for Wh or such, set to 3600."
        - name: integralInterpolation
          in: "query"
          type: "string"
          description: "(optional) when using aggregate method integral defines the interpolation method (defaults to none)."
          enum: [linear, none]
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HistoryResponse"
        422:
          description: "Invalid options supplied"
  /addHistory:
    x-swagger-router-controller: history
    post:
      tags:
        - "history"
      summary: "Add entry to history"
      operationId: "addHistoryByPost"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "Options"
          in: "body"
          description: "Options for add history"
          required: true
          schema:
            $ref: "#/definitions/HistoryAddRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HistoryAddResponse"
        422:
          description: "Invalid options supplied"
  /addHistory/{stateId}:
    x-swagger-router-controller: history
    get:
      tags:
        - "history"
      summary: "Add entry to history"
      operationId: "addHistoryByGet"
      produces:
        - "application/json"
      parameters:
        - name: "stateId"
          in: "path"
          description: "State ID"
          type: "string"
          required: true
        - name: val
          in: "query"
          type: "number"
          description: "Value"
          required: true
        - name: ts
          in: "query"
          type: "number"
          description: "Timestamp in ms. If empty current time is used"
        - name: ack
          in: "query"
          type: "number"
          description: "Acknowledge flag"
        - name: q
          in: "query"
          type: "number"
          description: "Quality flag"
        - name: from
          in: "query"
          type: "number"
          description: "From which instance"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HistoryAddResponse"
        422:
          description: "Invalid options supplied"
...
definitions:
...
  HistoryOptions:
    type: "object"
    #required: [pattern]
    properties:
      start:
        type: "number"
        description: "Start time in ms"
      end:
        type: "number"
        description: "End time in ms"
      count:
        type: "number"
        description: "number of values if aggregate is 'onchange' or number of intervals if other aggregate method. Count will be ignored if step is set, else default is 500 if not set"
      from:
        type: "boolean"
        description: "if FROM field should be included in answer"
      ack:
        type: "boolean"
        description: "if ACK field should be included in answer"
      q:
        type: "boolean"
        description: "if Q field should be included in answer"
      addId:
        type: "boolean"
        description: "if ID field should be included in answer"
      limit:
        type: "number"
        description: "do not return more entries than limit"
      ignoreNull:
        type: "boolean"
        description: "if null values should be included (false), replaced by last not null value (true) or replaced with 0 (0)"
      removeBorderValues:
        type: "boolean"
        description: "By default additional border values are returned to optimize charting. Set this option to true if this is not wanted (e.g. for script data processing)"
      returnNewestEntries:
        type: "boolean"
        description: "The returned data are always sorted by timestamp ascending. When using aggregate none and also providing count or limit this means that normally the oldest entries are returned (unless no start data is provided). Set this option to true to get the newest entries instead."
      aggregate:
        type: "string"
        description: "Aggregation method"
        enum: [minmax, max, min, average, total, count, percentile, quantile, integral, none]
      percentile:
        type: "number"
        description: "(optional) when using aggregate method percentile defines the percentile level (0..100)(defaults to 50)"
      quantile:
        type: "number"
        description: "(optional) when using aggregate method quantile defines the quantile level (0..1)(defaults to 0.5)"
      integralUnit:
        type: "string"
        description: "(optional) when using aggregate method integral defines the unit in seconds (defaults to 60s). e.g. to get integral in hours for Wh or such, set to 3600."
      integralInterpolation:
        type: "string"
        description: "(optional) when using aggregate method integral defines the interpolation method (defaults to none)."
        enum: [linear, none]
  HistoryRequest:
    type: "object"
    required: [id]
    properties:
      id:
        type: "string"
        description: "State ID"
      options:
        $ref: "#/definitions/HistoryOptions"
  HistoryEntry:
    type: "object"
    properties:
      val:
        description: "Value"
      ts:
        type: "number"
        description: "Timestamp in ms"
      ack:
        type: "boolean"
        description: "Acknowledge flag. Will be returned only if ACK set true in request options"
      from:
        type: "boolean"
        description: "Acknowledge flag. Will be returned only if FROM set true in request options"
      q:
        type: "boolean"
        description: "Acknowledge flag. Will be returned only if Q set true in request options"
      id:
        type: "string"
        description: "State ID. Will be returned only if addId set true in request options"
  HistoryResponse:
    type: "array"
    items:
      $ref: "#/definitions/HistoryEntry"
  HistoryAddEntry:
    type: "object"
    properties:
      val:
        description: "Value"
      ts:
        type: "number"
        description: "Timestamp in ms"
      ack:
        type: "boolean"
        description: "Acknowledge flag. Will be returned only if ACK set true in request options"
      from:
        type: "boolean"
        description: "Acknowledge flag. Will be returned only if FROM set true in request options"
      q:
        type: "boolean"
        description: "Acknowledge flag. Will be returned only if Q set true in request options"
  HistoryAddRequest:
    properties:
      id:
        type: "string"
        description: "State ID"
      state:
        $ref: "#/definitions/HistoryAddEntry"
  HistoryAddResponse:
    properties:
      success:
        type: "boolean"
        description: "true"
      error:
        type: "string"
        description: "Error description"
...